+++ TP1/matrix.c
@@ -4,7 +4,7 @@
 void check_fprint(FILE* fp, int copy);
 
 
-matrix_t* create_matrix(size_t rows, size_t cols) {
+matrix_t* create_matrix(unsigned int rows, unsigned int cols) {
     
     matrix_t* matrix = malloc(sizeof(matrix_t));
     if (!matrix) return NULL;
@@ -12,7 +12,7 @@ matrix_t* create_matrix(size_t rows, size_t cols) {
     matrix->rows = rows;
     matrix->cols = cols;
  
-    size_t n = rows*cols;
+    unsigned int n = rows*cols;
     matrix->array = malloc(sizeof(double)* n);
     if (!matrix->array) {
         free(matrix);
@@ -27,7 +27,7 @@ void destroy_matrix(matrix_t* m) {
     free(m);
 }
 
-int add_value(matrix_t* matrix_a, int *pos){
+int add_value(matrix_t* matrix_a, unsigned int *pos){
 	fprintf(stdout, "llego al add_value de matrix %d\n", pos[0]);
 	fprintf(stdout, "llego al add_value de matrix %d\n", pos[1]);
 	//falta guardar valor 1 en posicion pos[0],pos[1]
@@ -40,12 +40,12 @@ int print_matrix(FILE* fp, matrix_t* m) {
     size_t i = 0;
     int copy;
 
-    copy = fprintf(fp, "%ld", m->rows);
+    copy = fprintf(fp, "%d", m->rows);
     check_fprint(fp, copy);
 
     size_t n = m->rows;
     for (; i < n ; i++) {
-        copy = fprintf(fp," %f", m->array[i]);
+        copy = fprintf(fp," %d", m->array[i]);
 	    check_fprint(fp, copy);
 	}
